absa_yelp_project/requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
joblib>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0.1

# Sentiment + text
nltk>=3.8.1

# Tree models
xgboost>=2.0.0
catboost>=1.2.5

# Transformers
torch>=2.2.0
transformers>=4.41.0
datasets>=2.18.0
accelerate>=0.28.0

# IO
ujson>=5.9.0
pyarrow>=15.0.0

# Viz
matplotlib>=3.8.0
